backend/requirements.txt
fastapi==0.124.2
uvicorn[standard]==0.38.0
httpx==0.28.1
psutil==7.1.3
PyYAML==6.0.3
